requirements.txt
Django==1.6
dj-database-url==0.2.2
dj-static==0.0.5
django-annoying==0.7.9
django-crispy-forms==1.4.0
django-localflavor==1.0
django-toolbelt==0.0.1
gunicorn==18.0
httplib2==0.8
phonenumbers==5.9b1
psycopg2==2.5.1
pytz==2014.3
six==1.4.1
static==0.4
twilio==3.6.4
unittest2==0.5.1
wsgiref==0.1.2
